refactor(ai): remove leading underscore from OpenAI helper filenames
