+++ asyar-launcher/src/services/ai/providers/openai.ts
@@ -11,15 +11,15 @@ import type {
   ChatStreamEvent,
   ReasoningEffort,
 } from '../IProviderPlugin';
-import { buildOpenAIToolsBody, parseOpenAIToolStream } from './_openaiCompat';
-import type { OpenAIToolDescriptor } from './_openaiCompat';
+import { buildOpenAIToolsBody, parseOpenAIToolStream } from './openaiCompat';
+import type { OpenAIToolDescriptor } from './openaiCompat';
 import {
   buildOpenAIResponsesChatBody,
   buildOpenAIResponsesToolBody,
   parseOpenAIResponsesStream,
   parseOpenAIResponsesToolStream,
   usesOpenAIResponses,
-} from './_openaiResponses';
+} from './openaiResponses';
 
 const EFFORTS_BY_MODEL_FAMILY: Array<[families: string[], efforts: ReasoningEffort[]]> = [
   [
