+++ asyar-launcher/src/services/ai/providers/openaiCompat.ts
